+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- New function: `Get-FrontInbox`
+- Added more query filters to `Find-FrontConversation`
 
 ## [0.1.6] - 2021-11-22
 ### Added