+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.1.14] - 2021-12-15
 ### Added
 - New functions: `Get-FrontInbox`, `Get-FrontTag`, `Get-FrontMessageTemplate`, `Get-FrontMessageTemplateFolder`, `Update-FrontMessageTemplate`
 - Added more query filters to `Find-FrontConversation`
@@ -17,5 +19,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renamed parameter `-Tag` to `-TagId` in `Add-FrontConversationTag`
 - Use `PSCredential` object time to decode secure string instead of `ConvertFrom-SecureString` to consider older versions of PS
 
-[Unreleased]: https://github.com/codaamok/PSFront/compare/0.1.6..HEAD
+[Unreleased]: https://github.com/codaamok/PSFront/compare/0.1.14..HEAD
+[0.1.14]: https://github.com/codaamok/PSFront/compare/0.1.6..0.1.14
 [0.1.6]: https://github.com/codaamok/PSFront/tree/0.1.6